@@ -10,19 +10,19 @@
   },
   "dependencies": {
     "@types/cleave.js": "^1.4.6",
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.21",
-    "@types/react": "^17.0.39",
-    "@types/react-dom": "^17.0.13",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
     "@types/react-transition-group": "^4.4.4",
     "cleave.js": "^1.6.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-hook-useobserver": "^1.0.0-alpha.16",
     "react-icons": "^4.3.1",
-    "react-scripts": "^4.0.3",
+    "react-scripts": "^5.0.1",
     "react-transition-group": "^4.4.2",
-    "typescript": "^4.6.2"
+    "typescript": "^4.6.4"
   },
   "browserslist": {
     "production": [
