@@ -1,7 +1,7 @@
-import Vertical from "../../../layout/Vertical";
+import Vertical from "../../layout/Vertical";
 import React from "react";
-import {HeaderTitle} from "../../component/HeaderTitle";
-import {HeaderPanel} from "../../component/HeaderPanel";
+import {HeaderTitle} from "../component/HeaderTitle";
+import {HeaderPanel} from "../component/HeaderPanel";
 
 export function SettingsPanel(){
     return <Vertical style={{width:'100%',height:'100%'}}>
